Update Windows-Install-Binary.rst
I recently installed ROS2 on Windows 11 and found a couple of areas in which the documentation was unclear and am proposing the following changes in order to clarify.  Specifically, I proposed a change to note that it may work on Windows 11 (it's just not officially supported) especially since Windows 10 hits end of life in one month, and many people will want to try it on Windows 11, but the current documentation implies it will only work on Windows 10. Furthermore, making it clearer that you may need to move the files to the hardcoded path (I found they did not automatically install there even when I was in the directory and ran the install.)  I also noted that even though installation is done with Powershell that the actual running of ROS2 environment variable batch files fails in Powershell (Powershell lacks the 'CALL' command and gives a very confusing error if you try it) and that the normal CMD command prompt is needed instead.  (There is a workaround, but I think it would be more confusing than just saying to use CMD, so I'm proposing that change.)  Finally, the talker/listener seem to run better (at least in my tests) if you open the listener first, so I switched the order of the commands. Hopefully these changes help to make the install process easier for others!

Signed-off-by: James Sheldon <jsheldon@jamessheldon.com>